drain-to-feat for vehicle-types — relocate to bulletproof-react sub-areas
Pure relocation, no behavior change (ADR 2026-05-28).

- Move flat files into api/ components/{cells} hooks/ types/ utils/.
- pending-move/ parks the route-based @entur/vtype-details editor
  (VehicleTypeEditor.tsx + useVehicleType + fetchVehicleType), still
  served by /vehicle-types/new.
- Fix cross-feature DeckPlan/Name imports (deck-plans/*, useDataViewSearch)
  and App.tsx paths to the new locations.
