+++ src/App.tsx
@@ -8,8 +8,8 @@ import { useAppTheme } from './hooks/useAppTheme';
 import { ProtectedRoute } from './components/auth/ProtectedRoute';
 import { EditingProvider } from './contexts/EditingContext.tsx';
 import SessionExpiredDialog from './components/dialogs/SessionExpiredDialog.tsx';
-import VehicleTypeView from './data/vehicle-types/VehicleTypeView.tsx';
-import VehicleTypeDetails from './data/vehicle-types/VehicleTypeDetails.tsx';
+import VehicleTypeView from './data/vehicle-types/components/VehicleTypeView.tsx';
+import VehicleTypeDetails from './data/vehicle-types/pending-move/VehicleTypeEditor.tsx';
 import DeckPlanView from './data/deck-plans/DeckPlanView.tsx';
 import DeckPlanDetailsView from './data/deck-plans/DeckPlanDetailsView.tsx';
 import VehicleView from './data/vehicles/components/VehicleView.tsx';
