+++ src/data/deck-plans/deckPlanTypes.ts
@@ -1,4 +1,4 @@
-import type { DeckPlan } from '../vehicle-types/vehicleTypeTypes';
+import type { DeckPlan } from '../vehicle-types/types/vehicleTypeTypes';
 
 export type DeckPlanContext = {
   deckPlans: DeckPlan[];
