+++ src/data/deck-plans/fetchDeckPlans.ts
@@ -1,7 +1,7 @@
 import type { AccessToken } from '../../auth/index.ts';
 import { fetchDeckPlansRequest } from '../../graphql/vehicles/queries/fetchDeckPlans.ts';
 import type { DeckPlanContext } from './deckPlanTypes.ts';
-import type { DeckPlan, Name } from '../vehicle-types/vehicleTypeTypes.ts';
+import type { DeckPlan, Name } from '../vehicle-types/types/vehicleTypeTypes.ts';
 import type { Page } from '../../graphql/paginationTypes.ts';
 import { FETCH_ALL_SIZE } from '../../graphql/paginationTypes.ts';
