+++ src/data/vehicle-types/components/VehicleTypeView.tsx
@@ -1,9 +1,9 @@
 import { vehicleTypeViewConfig } from './vehicleTypeViewConfig.tsx';
-import GenericDataViewPage from '../../pages/GenericDataViewPage.tsx';
-import type { VehicleType } from './vehicleTypeTypes.ts';
-import type { OrderBy } from './useVehicleTypes.ts';
-import AutosysImportFloatingMenu from '../vehicle-imports/components/AutosysImportFloatingMenu.tsx';
-import { useUrlFilters } from '../../hooks/useUrlFilters.ts';
+import GenericDataViewPage from '../../../pages/GenericDataViewPage.tsx';
+import type { VehicleType } from '../types/vehicleTypeTypes.ts';
+import type { OrderBy } from '../hooks/useVehicleTypes.ts';
+import AutosysImportFloatingMenu from '../../vehicle-imports/components/AutosysImportFloatingMenu.tsx';
+import { useUrlFilters } from '../../../hooks/useUrlFilters.ts';
 
 export default function VehicleTypeView() {
   const { hasUrlFilters, clearUrlFilters, filterCount } = useUrlFilters();
